Fix location issue

+++ app/src/main/java/com/example/canooweather/data/entity/CurrentEntity.kt
@@ -21,7 +21,7 @@ data class CurrentEntity (
         @SerializedName("dew_point")
         var dewPoint: Float,
         @SerializedName("uvi")
-        var uvi: Int,
+        var uvi: Float,
         @SerializedName("clouds")
         var clouds: Int,
         @SerializedName("visibility")
